Reworked closures, per [Contributor 7746].
This spreads the burden of examples out a bit more -- the second example
doesn't need to show you every single possible shorthand all at once. It
also makes more sense, as he pointed out, to introduce the $0 form in
the context of a closure passed to a function, because that's where
you're likely to actually use it.