Adjusted discussion of binary expressions.
Although the formal grammar creates a tree with the operands associated
with the operator before it, the abstract syntax tree used in the
compiler just makes a flat list of operands and operators.  Feedback
from [Contributor 6004] was that we should avoid giving the impression that the
these things start off somehow left associative.