Spelling fixes.
Was running spell check on the Guided Tour, and figured I may as well
get the rest of the book while I was at it.

APSG says to use "indexes" rather than "indices" except when referring
to mathematical indices.  Perhaps one could make the argument that the
indexes into an array are mathematical in nature, but choosing the
regular plural seems simpler for ESL readers.