Wording tweak for Closures to satisfy some web feedback pedantry.
The feedback giver has a fair point…
In the "Closure Expression" section you say "Closure expressions provide several syntax optimizations for writing closures in their simplest form without loss of clarity or intent."
But in the "Capturing values" section you say "The simplest form of a closure in Swift is a nested function".
They cannot both be in the simplest form so one of these is wrong.