Clean up some failing tests.
Removed the mention of iterating over the characters of a string
because, for Unicode-correct results, you should typically not be
actually iterating over a string character by character.  So it's not
something we want to actively point people towards.

Left the puzzle input example because the problem space of making a
trivial little puzzle actually does make sense to operate over
characters, and the String.CharacterView struct is giving you extended
grapheme clusters, which are what you'd need for correctness.