Use descriptive placeholder type in Array examples.
This commit is a partial revert of 77af9e9e.

Original motivation from the editorial pass was to make this example
match the declaration of the type in the Array reference.  The
declaration has changed to Array<Element> as part of a pass over the
standard library to use descriptive names, so any benefit we might have
had from the opaque type placeholder is no longer there.  The original
SomeType carries the clear suggestion to the reader that they are
expected to replace SomeType with some type in their code.