Use print() instead of ephemeral variables.
The fact that these variables are created in the scope of the case is
causing confusion - I've seen several bits of dev feedback about it.
For example, feedback 283736.  The reason for assigning it to a variable
was because previous Xcode versions gave a warning about bare
expressions and those also confused developers.

As an alternative solution, I could have written "let vegetableComment:
String" outside the switch and assigned it a value in each case, but
that behavior seems more confusing.  The guided tour doesn't really
point out the fact that you can initialize a constant in a branching
structure like that.