Fix missing word.
See feedback ebb2a53b.

This prose could probably do with another polish pass and a better term
instead of "scope".  It's awkward because we don't actually have such a
thing as a "try expression" in our grammar: the "try" operator is just
allowed at various points in the expression grammar.