Move #available to a more sensible place.
Discussed with Brian in person; the old location looks like a mistake.
There isn't a perfect place to move this to, but making it a top-level
head at the end is the least bad change.  That also puts it near the
conceptually-related material about #if.

Fixes <rdar://problem/23868307> TSPL: Reference Having #available under
Control Transfer is confusing.