Remove a var pattern from guard.
This is more explicit than before about the fact that we are updating an
item struct with the new "count" value and then writing it back to the
array.  Before it looked sort of like, since it was a variable, it might
be getting written back by virtue of the modification, but in reality
that only triggers copy-on-write for the new item struct.

Brian reviewed the change in person.