Revise Bank example per [Contributor 9502]'s and Brian's feedback.
Follow-up polish for the first renaming in b0422eee1f.  Brian and I
considered "pay out", but it's pretty colloquial and doesn't appear in
several of the dictionaries we checked.  [Contributor 9502]'s suggestion of
"distribute" carries the right meaning of one central subject (the bank)
giving to several different objects (the players).