Remove discussion of mutating/nonmutaing pairs.
Method pairs like sort() and sorted() used to be related to
@warn_unused_result because you could use that attribute to attach
additional diagnostic information to them, which would be shown when you
probably used the wrong one -- as detected by the fact that you ignored
a return value.

Now that the polarity has been flipped, there's no conceptual link
between such method pairs and @discardableResult.  It's methods that
*don't* have discardable results where you want to add the cross
references.