Resolve the remaining pageturn fixmes.
For the arity fixme, I left the text as-is. It's in a section about using more
than one parameter in a function. Spending time talking about how functions are
disambiguated (additionally) by the text of their labels seems out of place
given 1) the current subsection and 2) the example is only an arity difference.

The as-is text is correct and the reasoning is quite relevant to the topic at hand.

I followed the suggestion for the other fixme. It reads great in the new location.

Fixes <rdar://problem/26903328> TSPL: Fix issue from WWDC pageturn