Use explicit dot syntax for enum cases.
SE-0036 makes the old syntax a warning.  I'm surprised we had it without
the dots to begin with, since one of the arguments for the SE proposal
was that the undotted syntax was very rare among developers and many
folks outside the compiler team didn't even know that it worked.