Reorder Any before AnyObject.
With id being imported as Any, it's clear that Any is the much more
relevant type for most work.  The only time you'd use AnyObject is for
something like a reference to a delegate object, where it's part of the
API contract you're making that the delegate has to be an instance of a
class, because the delegate has to have reference semantics.