Give more context for the listing.
The function names are long and looks similar -- it's easier to approach
the example if you already know what you're reading for.  Switch the
order of the calls to match the order the sentences need to go in and
the order of the declarations, instead of calling the most
recently-defined one first.

Per review with Brian.