Finish pass of AdvancedOperators and incorporate Brian's feedback
Addresses <rdar://problem/27597958> TSPL: SE-0077 Improved operator declarations

Finishes off a pass that removes all of the old syntax and introduces the new
syntax for custom operators and precedence groups.

The discussion of associativity in AdvancedOperators seems off, now, because
the example wants to have the same behavior as addition. But since that's now
defined by AdditionPrecedence, there's not necessarily a reason to spell out
how you'd otherwise set it to something custom.

Stylistic and typo fixes from Brian.