Add missing production for a labeled operator.
The version of this grammar on "develop" covered this case, but it
wasn't covered at the point this branch diverged.  This 2x2 grid of
alternate productions isn't our usual shape, but it has the correct
semantic grouping; we liked it better that all on one line (hard to
read, not our usual style) or four lines (loss of semantic information).

Review with Brian and Kyle.