Add brief mention of methods.
Didn't add anything to the example, because it doesn't fill well, and
it's not actually needed by any of the later listings in this chapter,
unlike properties which actually are.  At this point, the reader has
probably seen methods in the Tour and already got a cross reference to
the relevant chapter in the bulleted list up top.