Clean up test code.
The REPL can't seem to parse """ strings in non-compiled mode.

Starting one line with a " is useful, but starting most of them that way
could be confusing -- especially because during the SE list bikeshedding
for this feature, there was a suggested syntax where you started each
line of a muliline string with " to mark it as such.