Fix the link to String.
This link used to work, back when the stdlib ref was with all the other
reference in [Tool M], in the pre-[Tool B] world.  The build system doesn't
try to fix this, since [Tool J] doesn't know about [Tool B].  It looks like
this was the only example of a link to a reference doc in this book.