Add grammar for escaped newline, and for whitespace.
Adding grammar for escaped newline prompted the need for adding grammar
for whitespace, which prompted the need to add grammar for comments.
Since this is still the only place where the grammar refers to
whitespace, if we can do it differently, we won't need to add the other
grammar-of box.

Not entirely happy with the production that disallows /* and */ inside
of a multiline comment.