Fix ordering issue.
Concatenation is introduced later in the chapter, so don't use it as
part of the explanation of multiline string literals.  Bring back the
content from swift-4-beta-1 as part of the original intro, and move the
new example down to be a practical tip around concatenation.

In-person feedback from Brian and Kyle.