Reframe rule about closure parameters and change the example.
The rule is easier to understand if you think of it as
a rule about all nonescaping closure parameters.
You simply can't pass them on as arguments to other
functions that also came in as closure parameters.

The new example, while less practical, better shows the shape
of what's not allowed.
It shows several variations to give the hint that
no matter how you arrange the demoed calls,
you're going to run into this issue.