Refer to closures as function types in the Function Types section
Because I'm in the grammar for function types,
it doesn't really make sense to use the word 'closure' a bunch.

Even though people conceptually think of these things as closures
sometimes, the rule is about types, and this is a section about types.

Plus, my example doesn't actually have any closures in it at all.