Remove a fix that doesn't work.
The way you'd fix this here is to make a local copy and use that on
one/both sides of the balance() call.  There's not much point going
through that example though, because we already have a discussion above
that shows making an explicit copy.

I'm not sure why I thought it worked, but the doubts in my TR query are
essentially why it doesn't: although the calls to the new
function/method don't have overlapping accesses, their implementation
still does.