Start applying hardcopy edit.
Don't try to explain why lines marked "// Error" have an error in them.
In all cases, it's because there's a conflict that's explained in the
paragraph immediately after the code listing.  I initially tried the
opposite -- adding "because foo and bar conflict" after each one -- to
match what's done in the other place in the guide where we have "Error"
comments.  Explaining why there's an error requires more information
that I can fit in a <40 character inline comment.