Finish framing classes vs structs.
Per discussion with Kyle, the new article covers the details of
why structs are easier to reason about than classes, although I point it
out briefly after the class figure.  Dropped the examples of what makes
a good struct/class because that discussion belongs to the article.

Explanation of COW is in a note because readers could legitimately skip
it if they don't care that much about performance.  It's at the end of
the section because slotting it in before the example starts breaks the
flow too much and inverts the most-to-least-salient structure.

Drop mention of NSArray and friends being classes.  That was useful in
the Swift 1 era because it was unexpected to Obj-C programmers, but it's
no longer needed.