Make the COW discussion more hand-wavey.
Kyle did some digging here, and in most cases it's not actually
copy-on-write behavior, it's that the optimizer has determined that it
doesn't need to make a copy (which may be because the custom struct's
properties all implement COW).

https://forums.swift.org/t/implementation-of-swifts-value-types/3404/2
Plain struct types are not copy on write. But really create a copy.

https://forums.swift.org/t/passing-value-types-or-members-of-value-types/5869/2
Array, and other types, implement copy-on-write themselves, but this is
not true of arbitrary structs.