Use a fake AppleRef to trigger a link.
The link still isn't available in [Tool S], likely because we're no longer
turning the cranks on that system, so the real AppleRef that would be
needed by [Tool G]/[Tool J] isn't available.  That's actually not a problem
though -- we're done building with those too.  We'll add a URL to the
AppleRef translation table that the web_transition branch created.