Add 'less than' to platform condition reference.
I tried a few different formulations of the grammar,
this one looked best. In particular, separating out the operators
into their own production looked odd.

I thought about bumping the version examples to Swift 5,
but do we really want to be bumping that number in perpetuity?

I used Swift 3 (with no .0) to add some flavor to the examples
using the new operator.