Include the new chapter in the "part" chapter.
It needs to be added both here and in "[Tool S] Chapters".  This one makes
it show up in the "toctree" for Sphinx when building HTML.  The other
one gives the chapter its ID and makes it show up when building [Tool S]
XML, which the makefile still does for historical reasons.