Rough in memberwise init with default values.
I don't see a place in the reference that needs to be updated for this
change.  We don't define the initialization rules formally in the
reference, so there isn't anything there that would now be incorrect.
Memberwise initializers don't have a corresponding declaration because
the whole point of them is that you *don't* have to write one.