Consistently call these "opaque types".
We don't use the term "result type" in TSPL, so it doesn't make sense to
talk about the "opaque" kind of "result type".  If we wanted another
word in there, we'd probably call it an "opaque return type".  The SE
proposal uses both "opaque type" and "opaque return type" throughout,
as well as "opaque result type".