Be more specific.
One line of code isn't quite the same as one statement.

Alternate, rejected rewrite:

If the body of a function is only
the ``return`` keyword and one expression,
you can omit the ``return`` keyword
and write just the expression.
This implicit ``return`` syntax lets you write
very simple functions in a shorter way.