Correct the name of the property wrapper.
It actually allows the number 12 too, which was overlooked the first
time it got renamed.  I started taking a pass to change the twelves to
elevens and make the behavior match the current name, but renaming it is
much easier and less error prone.

Per review with Brian and Kyle.