Standardize naming for do-catch.
Following on from discussion with Chuck this afternoon.  We use "clause"
elsewhere to refer to just the 'do' or 'catch' part, but not to refer to
a complete statement.  "Block" is also technically correct -- or, at
least, it's correct to say that a do-catch statement contains two code
blocks -- but we don't really need to use two different terms.

The guide could potentially call it a "block" if we want it to avoid
saying "statement" in a future revision.