Remove straggler ==> in grammar.
The old list of transformation to turn BNF into TSPL-style grammar used
that notation, but that was dropped in favor of describing the TSPL
grammar on its own.  The old rules came from writing down how Brian and
I translated the BNF from the langref, but these days there's a lot less
BNF input to be translated, so expressing the rules that way is less
useful.