Correct the example.
In a C function, you'd take a pointer to a null-terminated C string, not
a pointer to a Swift string.  So a realistic example of that would have
a parameter of type UnsafePointer<CChar> instead and use the implicit
conversion from String to that type -- which isn't marked by & at the
call site.

Per tech review from Karoy Lorentey <klorentey@apple.com>.