Make the "ignore this" underscore visible.
When we put the hidden underscore line in, back in commit 06cd158b2,
this is the alternate approach that seemed like it would be too hard to
approach.  Calling it out in the experiment box should help folks follow
along with this syntax -- we don't show using underscore to ignore a
value anywhere explicitly in the tour.