Update figure to correctly show "nil".
Jerimiah also removed the gap in the gray line -- it makes the PNG look
weird because of how the white and transparent bits interact.  There
aren't any figures like this elsewhere in the book that we need to
match.  Diagrams like CollectionTypes_intro_2x.png have overlapping
black arrows, and this is the only diagram in the ARC chapter (where we
use colored arrows) that has overlapping lines.

Asset from <rdar://problem/74529003>.